check routes authorizations in controller

+++ src/Controller/UsersController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Entity\User;
+use App\Entity\Client;
 use App\Repository\UserRepository;
 use App\Repository\ClientRepository;
 use Doctrine\ORM\EntityManagerInterface;
@@ -11,6 +12,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Serializer\SerializerInterface;
+use Symfony\Component\Security\Http\Attribute\CurrentUser;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
@@ -30,7 +32,7 @@ class UsersController extends AbstractController
     }
 
     #[Route('/', name:"createUser", methods: ['POST'])]
-    public function createUser(Request $request, SerializerInterface $serializer, EntityManagerInterface $em, UrlGeneratorInterface $urlGenerator, ClientRepository $clientRepository, ValidatorInterface $validator): JsonResponse
+    public function createUser(#[CurrentUser] ?Client $connectedClient, Request $request, SerializerInterface $serializer, EntityManagerInterface $em, UrlGeneratorInterface $urlGenerator, ClientRepository $clientRepository, ValidatorInterface $validator): JsonResponse
     {
 
         $user = $serializer->deserialize($request->getContent(), User::class, 'json');
@@ -42,11 +44,8 @@ class UsersController extends AbstractController
         return new JsonResponse($serializer->serialize($errors, 'json'), JsonResponse::HTTP_BAD_REQUEST, [], true);
         }
 
-        // Récupération de l'ensemble des données envoyées sous forme de tableau
-        $content = $request->toArray();
-
         // Récupération du client_id.
-        $clientId = $content['client_id'];
+        $clientId = $connectedClient->getId();
 
         // On cherche le client qui correspond et on l'assigne à l'utilisateur.
         // Si "find" ne trouve pas le client, alors null sera retourné.
@@ -63,25 +62,34 @@ class UsersController extends AbstractController
    }
 
    #[Route('/{id}', name: 'detailUser', methods: ['GET'])]
-    public function getDetailUser(User $user, SerializerInterface $serializer, UserRepository $userRepository): JsonResponse {
+    public function getDetailUser(#[CurrentUser] ?Client $connectedClient, User $user, SerializerInterface $serializer): JsonResponse {
 
+        if($connectedClient == $user->getClient()){
         $jsonUser = $serializer->serialize($user, 'json', ['groups' => 'getUsers']);
         return new JsonResponse($jsonUser, Response::HTTP_OK, ['accept' => 'json'], true);
+        }else{
+            return new JsonResponse('You don\'t have rights to access this page');
+        }
    }
 
    #[Route('/{id}', name: 'deleteUser', methods: ['DELETE'])]
-    public function deleteUser(User $user, EntityManagerInterface $em): JsonResponse
+    public function deleteUser(#[CurrentUser] ?Client $connectedClient, User $user, EntityManagerInterface $em): JsonResponse
     {
+        if($connectedClient == $user->getClient()){
         $em->remove($user);
         $em->flush();
 
         return new JsonResponse(null, Response::HTTP_NO_CONTENT);
+        }else{
+            return new JsonResponse('You don\'t have rights to delete this user');
+        }
     }
 
    #[Route('/{id}', name:"updateUser", methods:['PUT'])]
-   public function updateUser(Request $request, SerializerInterface $serializer, User $currentUser, EntityManagerInterface $em, ClientRepository $clientRepository, ValidatorInterface $validator): JsonResponse
+   public function updateUser(#[CurrentUser] ?Client $connectedClient, Request $request, SerializerInterface $serializer, User $currentUser, EntityManagerInterface $em, ClientRepository $clientRepository, ValidatorInterface $validator, UrlGeneratorInterface $urlGenerator): JsonResponse
    {
-       $updatedUser = $serializer->deserialize($request->getContent(),
+    if($connectedClient == $currentUser->getClient()){
+        $updatedUser = $serializer->deserialize($request->getContent(),
                User::class,
                'json',
                [AbstractNormalizer::OBJECT_TO_POPULATE => $currentUser]);
@@ -94,12 +102,20 @@ class UsersController extends AbstractController
        }
 
        $content = $request->toArray();
-       $clientId = $content['client_id'];
+       $clientId = $connectedClient->getId();
        $updatedUser->setClient($clientRepository->find($clientId));
 
        $em->persist($updatedUser);
        $em->flush();
-       return new JsonResponse(null, JsonResponse::HTTP_NO_CONTENT);
+
+       $jsonUser = $serializer->serialize($updatedUser, 'json', ['groups' => 'getUsers']);
+
+       $location = $urlGenerator->generate('detailUser', ['id' => $updatedUser->getId()], UrlGeneratorInterface::ABSOLUTE_URL);
+
+       return new JsonResponse($jsonUser, Response::HTTP_OK, ["Location" => $location], true);
+    }else{
+        return new JsonResponse('You don\'t have rights to update this user');
+    }
   }
 }
